Fix character module unit tests
