+++ default/python/tests/AI/character_test/character_strings_test.py
