+++ default/python/tests/AI/character_test/character_test.py
